Fix move commonMiddlewareConfig.js to middlewares folder
