+++ Server/Middlewares/commonMiddleware.js
@@ -3,7 +3,7 @@ import morgan from "morgan";
 import express from "express";
 import cors from "cors";
 import cookieParser from "cookie-parser";
-import { NODE_ENV, CORS_ORIGIN } from "./envConfig.js";
+import { NODE_ENV, CORS_ORIGIN } from "../Config/envConfig.js";
 
 export function applyCommonMiddleware(app) {
   // ---- Middlewares (safe defaults) ----
